pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bwip"
version = "2.0.0"
description = "Bathing Water Information Portal - Generate bathing water information posters for Irish Local Authorities"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.11"
authors = [
    {name = "BWIP Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Django core
    "Django>=6.0,<7.0",
    "psycopg[binary]>=3.1",

    # Authentication with MFA
    "django-allauth[mfa]>=65.0",

    # Admin theme
    "django-jazzmin>=3.0",

    # REST API
    "djangorestframework>=3.14",

    # PDF generation
    "weasyprint>=60.0",
    "qrcode[pil]>=7.4",
    "Pillow>=10.0",

    # HTTP client
    "requests>=2.31",
    "httpx>=0.25",

    # Data validation
    "pydantic>=2.5",

    # Environment management
    "python-decouple>=3.8",
    "dj-database-url>=2.1",

    # Utilities
    "django-extensions>=3.2",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4",
    "pytest-django>=4.5",
    "pytest-cov>=4.1",
    "factory-boy>=3.3",
    "faker>=20.0",

    # Code quality
    "ruff>=0.1",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.7",
    "django-stubs>=4.2",

    # Development utilities
    "django-debug-toolbar>=4.2",
    "ipython>=8.0",
]

[project.urls]
"Homepage" = "https://github.com/your-org/bwip"
"Documentation" = "https://github.com/your-org/bwip/docs"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "services*", "config*", "tasks*"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.isort]
known-first-party = ["apps", "services", "config", "tasks"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.black]
line-length = 88
target-version = ["py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["apps", "services", "config", "tasks"]
skip = ["migrations", ".venv"]

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
strict = false
warn_return_any = true
warn_unused_configs = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "--strict-markers",
    "-ra",
    "-q",
]
testpaths = ["tests", "apps"]

[tool.coverage.run]
branch = true
source = ["apps", "services"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
